[dirmd] Add support for subst
subst is a Windows-specific command that maps a drive to a path on
another drive, e.g. it can map D: to C:\chromium\src.
Currently dirmd works incorrectly when executed in the context of
subst-ed paths and this CL fixes that.

If a path is rooted in a subst-ed drive, then most dirmd subcommands
replace them with the canonical path, e.g. D:\foo ->

C:\chromium\src\foo. The only exceptions are deprecated subcommands as
a nudge to use their replacements(`export` and `compute` were replaced
with `read`), as well as subcommands that very rarely executed by
consumers.

Bug: 1223076
Change-Id: Ic28e15cfd1616b537db711edf6cf15f0d05771e1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3192331
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44188}