[Weetbix] Add thresholds to project configuration.
Allow projects to specify bug-filing thresholds and thresholds for
applying priority values to bugs. In a follow-up change (and once
the configuration has been set), the configuration will be used.

Validation is not backwards-compatible but this is not a concern
as we are not yet in prod.

BUG=chromium:1243174
TEST=unit tests

Change-Id: If2a9ac2f472e4119abd828fc0e63bba90acae5b9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3190131
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#44202}