setup_project: switch gs/gitiles clients to stop using OAuth
After our GCP project finished the 1p verification process,
I learned that OOB (the OAuth browser-login-copy-and-paste-secret
flow that luciauth packages rely on) is not supported for 1p
apps. Switching the setup_project app to production effectively
bricked the application, because OAuth no longer works and you
can't switch 1p apps out of production once they've been switched on.

I could create a new pantheon project with a new OAuth client secret
and go back to our temporary fix of publishing the app in "testing"
mode and manually whitelisting each partner email, but this is not
a long term soln -- I'd still need to implement a more complicated
OAuth scheme described in b/199185011 which is a lot of work and
would likely require changes to upstream luciauth packages.

Instead, I decided to change the way this (intended-to-be) lightweight
application uses gs and gerrit.

For GS, I shell out to `gsutil`, which partners should have in their
path.

For Gerrit/Gitiles, I created a new type of Gitiles client that leverages
the public github.com/andygrunwald/go-gerrit package that we use in
branch_util to create branches. This client supports authenticating
with git cookies, which is what I do -- I parse ~/.gitcookies for the
correct client secret, and pass that during initialization.

All in all, this will make this application lower touch for partners
(they won't have to log in) which I think is a win, even if the code
quality is slightly worse. The use of existing auth in the invoker's
environment is in line with what's currently done in setup_project.sh.

BUG=b:187795796
TEST=go test, manual runs

Change-Id: Ia2d564769f89ef8c657fa4ac75f5e98e122b5675
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3194758
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Cr-Commit-Position: refs/heads/main@{#44209}