[Findit] Enable v2 for all ci builders for compile failure
Findit v2 is only running for a subset of builders. To turn off v1, v2 needs to run for all ci builders. Turning off v1 also helps with gbeaty@'s proposal to integrate Findit with bootstrapper.

Although only run for a small subset of v1, in the month of August, v2 reverts more than v1 in August (out of 9 reverts submitted, 4 are of v1 - so the remaining 5 should be the work of v2 - this needs to be confirmed). So enabling v2 for all ci builders has the potential to improve Findit's impact as well.

One thing to note here is that this may increase the running cost - I will monitor the cost to make sure it does not increase too much.


Bug: 978503
Change-Id: I7b07d7fd677fb3b0102a58f34b0a67aadf542ba2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3189283
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#44258}