Change feature coverage queue from pull to push.
Pull queue tasks are currently handled at startup, which poses a problem that app engine has a hard deadline of 24 hours for any request. Since we run an indefinite loop in /_ah/start handler, it doesn't return at the end of 24 hours and instance is considered 'not-ready' and it errors out.
This is an issue because, we could have a feature coverage logic executing at 24 hour boundary, which means that in the case of instance being killed, only partial data would get exported to bigquery, affecting correctness of feature coverage dashboard.


Bug: 1141021
Change-Id: I1a868309a2e51fbcd655d65b7b24847614c58bbe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3199307
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Auto-Submit: Prakhar Asthana <pasthana@google.com>
Cr-Commit-Position: refs/heads/main@{#44265}