Remove lacros-amd64-generic-chrome-skylab from creating SoM alerts.
This is creating SoM alerts during skylab outages, and sheriffs have
no knowledge on how to triage the failures. It's a frequent source of
confusion for sheriffs.

Given how bespoke the builder is, and how it runs tests, it should
have a more focused/relevant sheriff rotation rather than the main
chromium sheriffs, who are overloaded as it is.

Bug: None
Change-Id: Iba9decffd1314795fb6825ebaad70194d84a3fd6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3198321
Reviewed-by: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44408}