[recovery]: ReadLine function to execute commands on hosts
This commit is a part of the PARIS project.

This change introduces a function-type that be used for executing
commands on different hosts, such as the servo-host, or the DUT.

This commit also includes a specific function that returns a function
of the above mentioned type.

This change will be used in subsequence CLs, such for servo root
verifier and servo topology verifier.

bug=b:198809511
TEST=None

Change-Id: Idd3a7289ebc69f4dc261dcbbe403bfc5c6e8e27a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3217951
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#44416}