upload_debug_symbols: create main driver for distributed worker design
Implement a distributed worker design to handle the debug symbol
uploading. This workflow spawns up to a user configured amount of
goroutines to upload the debug symbols.

A slight redesign was made for the upload function and worker
communication. First, the upload function was adjusted to not use
channels and instead just receive the task as an argument.

BUG=b:195981092,b:197010274
TEST=unit tests

Change-Id: Idabd6f625dbe364e148507191fbeb3b6d396e8fc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3165920
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Cr-Commit-Position: refs/heads/main@{#44583}