upload_debug_symbols: Generate body for batch duplicate check
Before we can upload a debug symbol we need to check if has already been
uploaded. Given that on any given run there could be thousands of
symbols we will perform a batch checking of symbols rather than checking
each individually.

In this change we are only generating the body of the request. Once the
https client is in place this generated body will be sent alongside it.

To verify the format of the body, I did a local run of this builder and
used the outputted body in a CLI curl call. This returned the expected
results verifying the body generation.

BUG=b:195981092,b:197010274
TEST=unit tests, local executions

Change-Id: I3ac75fb5f872158847fd092dd4563ee7dfbc8871
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3205973
Commit-Queue: Julio Hurtado <juahurta@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Cr-Commit-Position: refs/heads/main@{#44603}