[Weetbix] Verify keep-open threshold is satisfied by bug-filing thresh.
The "With multiple clusters above impact threshold" test was failing
because an opened bug was being immediately closed. This was because the
keep-open threshold was more onerous than the bug-filing threshold,
which is not a useful system configuration.

Update configuration validation to exclude this case and add an
assertion to the test that validates the configuration being used is valid.

BUG=1243174
TEST=unit tests

Change-Id: I95a153be5dae49af57676b6cd871a753ba2be8c1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3220776
Reviewed-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#44629}