[Weetbix] Hook clustering ingestion into main Weetbix ingestion.
Failures will now be ingested for clustering in addition to for
variant analysis.

This change adds a mandatory partition_time field to the
resultingester task proto, which will cause in-flight ingestion
tasks to fail. This was considered acceptable as we are not in
production.

BUG=chromium:1243174
TEST=unit tests

Change-Id: Ib613c497d6a844d784d03679215115e2db683ae0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3220109
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#44640}