Remove unused testing library and problematic dependency.
I'm looking to add `gopls` to tools.go, however gopls has a documented
warning when it's built with go-diff
(https://github.com/golang/go/issues/45648) and spits out errors when
it detects this library.

We only have this dependency for an (apparently) unused library, so
remove the library and the direct dependency.

Introduce a replace directive which pins us to 1.1.0.

R=vadimsh

Change-Id: Ie0315700ba8e7513c6105240437d4afc3df80aaa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3235587
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44652}