[WIB] Add support to write configs to disk in WScripts executor
Bug: b/204104098

Add support to write back pinned configs in windows scripts executor.
This config file will be used to estimate an unique identifier for the
image. This will also be copied to the image in root directory. Running
this config is expected to produce the same image again.

Change-Id: I4ed5f9a9eadba727f1f60d760772197e83d57343
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3199467
Commit-Queue: Anushruth  <anushruth@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44721}