[Findit] Only run v2 analysis if we can find a passing step
Currently Findit v2 runs compile analysis even when the step in the previous build was a failure. E.g.

https://ci.chromium.org/ui/p/chromium/builders/ci/ToTAndroidCoverage%20x86/1329/overview
and
https://ci.chromium.org/ui/p/chromium/builders/ci/ToTAndroidCoverage%20x86/1328/overview

This leads to a lot of false positive. We should only let Findit try to analyse if the previous build/step was successful.


Bug: 978503
Change-Id: Ic4ec38c0e282d5a271ea1ac185f05a9fd46b9999
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3238506
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#44776}