[Findit v2] For chromium, only proceed if we can find suspect to rerun
As discussed, to reduce the number of false positives for v2, we want to only take action when the result of heuristic analysis (i.e. suspects), and the result of bisection are the same (supposedly we only want to do this for chromium). This is also a condition for v1.

This effectively means that for rerun analysis, because we already have the results of heuristic analysis, we can just verify the result by running 2 rerun, 1 for the suspect, and 1 for the suspect's parent.

Bug: 978503
Change-Id: I5ae768b2f8c7bd814c8fd6fcf0450501dc445597
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3245054
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Cr-Commit-Position: refs/heads/main@{#44823}