Download the bootstrapped executable directly to CIPD root.
Currently the bootstrapper downloads the bootstrapped executable to a
path in the CIPD root with the same name as the package. This results in
very long path names, which caused the file names on windows to exceed
the path length limit. This changes it so that the bootstrapped
executable's package is downloaded directly to the CIPD root that is the
exe subdirectory in the build's start directory.

Bug: 1264242
Change-Id: I7e036114d21fc1bc81213b9f885fa1df8ba48314
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3252296
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#44852}