[recovery]: Function to parse the action args
This commit is a part of the PARIS project.

This change includes a function that splits the action args into key
and value pairs using a splitter string. This utility function will be
called whenever any exec function needs to parse the action args.

bug=b:198809511
TEST=build

Change-Id: Ie6215884ff7d8001cbe8266b2809bc6965e037ef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3256875
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#44914}