[Weetbix] Update bug-filing to create failure association rules.
This change revises automatic bug-filing to create failure association
rules for each bug, rather than simply mapping them to suggested
clusters. It also changes the source of analysis to be Weetbix's
internal analysis of clustered failures.

Once re-clustering is implemented, we will be able to close the
feedback loop, with analysis of bug clusters (as defined
by the failure association rules) driving bug priorities in
auto-bug filing. For now, we use analysis of the source suggested
cluster from which the failure association rule was created to
adjust bug priorities. This is similar to the existing logic.

BUG=1243174
TEST=unit tests

Change-Id: I76d9bda5b1dfa01e96bd036a76184d071bac084e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3258881
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#44973}