upload_debug_symbols: add in crash api connection
This change adds in http functionality to connect with the crash API.
Specific effort was made to follow the same process as the legacy
builder to ensure functionality remained the same.

Another key feature of this change was that logging was added in. For
now it's still fairly limited but it details the steps taken and
responses from the crash api endpoints.

Testing also got a fair bit of attention as a mocking system was put
inplace to test functions that spoke to crash over http requests.

BUG=b:195981092,b:197010274
TEST=go test -v, local binary executions

Change-Id: I93a8b260608024a6fef27f332d0d7e6f253740e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3205974
Reviewed-by: Jack Neus <jackneus@google.com>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Cr-Commit-Position: refs/heads/main@{#44985}