[WIB] Refactor customization processor in scripts executor
Bug: b/204104098

This change contains a major refactor for the windows scripts executor.
Base class customization is used to contain all the common
functionality for the customizations. offline winpe customization is
a class that inherits the customization class. The source managers are
also standardized to simplify reading the code and execution.

Change-Id: I3f4bd6d4f7aa7bbe430b2afb05612d5c66f5ebe4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3220113
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/main@{#44992}