Update the build's gitiles commit based on bootstrapping.
This will prevent bootstrapping from causing a failure for chromium
builders that build from tag commits. Builders that build tag commits
are triggered with a gitiles commit value that does not have an ID set,
but bootstrapping causes a specific revision value to be used. An
assertion error is raised when attempting to set the output commit in
bot_update. Setting the gitiles commit on the input causes it to instead
set an output commit that uses the input commit's ref because the input
commit's ID will match the revision that was requested.

Bug: 1265336
Change-Id: Ibe7dd25b3e75b4da432685c61b482575c4b8ac00
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3268861
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#45061}