[Weetbix] Add methods to read current rules version, and rules by delta.
To describe which failure association rules a chunk was clustered with,
the chunk metadata contains a RulesVersion. This is the timestamp of the
last rule update in the snapshot of rules that was used to cluster the
test results. Provide a method to read it.

Furthermore, to facilitate timely updates of rules to the instances, a
method to read only rules changed after a certain timestamp is provided.

These methods are used in a follow-up CL.

BUG=1243174
TEST=unit tests

Change-Id: I4b05752f256cbd389b29cd3afa4025548e08c931
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3266573
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#45083}