[Weetbix] Refactor ClusteringState.
To better support incremental re-clustering in future, this change
explicitly captures all clustering algorithms used to cluster a
chunk as part of the chunk clustering state (not just the
AlgorithmsVersion). Although each AlgorithmsVersion technically
maps to a set of algorithms, we would rather not store a mapping
table from a version to a set of algorithms.

This change also wraps up the clustering of a chunk and its
associated metadata in one struct, which will again simplify
things when we start to implement re-clustering, as we can
modify the signature of Cluster() to be:
Cluster(existing ClusterResults, failures []*Failure) ClusterResults

BUG=1243174
TEST=unit tests

Change-Id: I965a01bdbe12b3057e40745ff5f772a74fba3a97
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3266575
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#45086}