Weetbix: add cluster failures API handler
This handler returns the last 2000 failures for a cluster.  This
information will be used to add data to the cluster page, both linking
to example failures and allowing multi-dimensional analysis (in browser
memory) over the latest failures.

2000 failures works out to ~1MB of data uncompressed in JSON transferred
to the browser.  Adding compression (or using a more compact format)
will enable transferring more data for analysis (10 000 faliures is
enough for 7 days of almost all of the known clusters).

Bug: 1243174
Change-Id: I6cb0c3e459c0e8ae9970b6446552fea966105f4b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3266583
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45152}