UFS: implement Convert library for FlatConfigs to Swarming labels
For a queried DUT, there is a FlatConfigs corresponding to the DUT.
Using the FlatConfig and predefined DutAttributes, the field path is
used to determine the location of the actual value in FlatConfig. Using
jsonpath library, the values of a particular label can be accessed
given a FieldSpec path.

The labels generated are in the form `${label-prefix}:val1,val2`, where
the values are comma-separated.

BUG=b/185639563
TEST=go test ./...

Change-Id: I404daf23e3143b4c20216923d84f7db291f692e2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3198779
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#45156}