SoM: Add parallel typescript/lit2 build
Existing code is built with polymer2, but this adds the ability to add
new components written in lit2 with type checking by typescript.

Added an experimental som-table-view component which displays alerts in
a compact table rather than the expansive current alert view.  This view
is very basic in this CL, and is probably not useful yet.

The new view can be reached by appending /alerttable to the url, e.g.
/chromeos/alerttable

Bug: None
Change-Id: Ia01e7793445c495def824787b9621466d6ed23cc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3266582
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#45162}