[recovery]: Determine whether the DUT is connected to servo using the 'ppdut5_mv' servod control.
This commit is a part of the PARIS project.

If the voltage value obtained from 'ppdut5_mv' servod control is
smaller than a threshold, it indicates that the DUT is not connected
to servo. This commit utilizes this check to verify that the DUT is
properly connected to the servo device.

Bug=b:198809511
TEST=build

Change-Id: Ib8576f3cb332c698f5a49dfde7968dfdb88e4fdb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3271813
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45283}