[karte] Use hex-coded IDs that count down.
Use hex-coded IDs that count down from the end of time instead of IDs that count up.
This is the approach used by, for example, swarming.

The first four letters of datastore IDs are reserved for the version of the ID forever
and the starting id is zzzz, this gives us (26 + 10 + 26)^4 (which is 14776336) possible values.

The latest records are lexicographically smallest. This means that queries that do not specify
any sort order will automatically get the most recent records first.

The resulting IDs are much more compact than the previous version.

BUG=b/205162975
TEST=make test

Change-Id: I414fc833f1634dbbbfe6486c85a6202c88f5bcaa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3277243
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45306}