[recovery]: Using warm_reset, detect if DUT is connected to servo.
This commit is a part of the PARIS project.

For setup that include servo type other than V3, the DUT may be
disconnected from the servo. The servod command warm_reset can be used
to detect such situations. This commit includes a verifier that uses
warm_reset command to detect if DUT is connected.

The warm_reset control is not essential to be supported by servod, and
if a servod does not support it, it is not an error. However, if
warm_reset is supported, it should return the correct value for
verification to succeed.

bug=b:198809511
TEST=build, manual test

Change-Id: Ibe649497babfd672cd24260b8d4df3ede5599abc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3286684
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#45374}