[Recovery] Set the log dest to stderr
Set the logging destination to stderr inside the recovery library. We
know from b:202156419 comment #5 that writing the stderr of the labpack
process will result in a message that is drained correctly, but is not
separated by step or level.

Also add some probes (a UUID in a log message) for the different log
levels so we can get more information about how the Go standard library
logger is configured by default in the environment where labpack runs.

BUG=b:202156419
TEST=make test

Change-Id: I8e8aac1036674bf893f854c9b2bb8004787ce7b3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3291562
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45408}