[labpack] Get infra proto in labpack main.go
This CL depends on this other CL over here that implements the Infra
method.
https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3292049

As described in the bug, we need information about the infrastructure,
specifically the buildbucket ID and the swarming task ID. This
information is available to luciexes and is present in the initial proto
that the luciexe framework processes, but is not exposed to the
user code until the linked CL lands.

BUG=b:202156419
TEST=make test

Change-Id: I09ace3988c554c5a980627663308dc524838eb8c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3291853
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#45433}