[recovery]: Check whether servod supports a control
This commit is a part of the PARIS project.

The servod daemon supports a range of controls. This commit adds a
verifier that queries servod about whether or not it supports a given
control.

This verifier is a building block for multiple steps in servo
verification.

Bug=b:198809511
TEST=build

Change-Id: I2964ef41fc005e4660ccb46b9145c3e770106b70
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3296003
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#45466}