[recovery]: Check if the ccd sbu voltage is floating between high and low values.
This commit is a part of the PARIS project.

The checks for whether CR50 is enumerated, and whether the CR50
voltage is low, are conditioned on whether the value of servod control
'dut_sbu_voltage_float_fault' is 'on'. This commit introduces a
verfier that checks the value of this servod control.

Bug=b:198809511
TEST=build, manual test

Change-Id: Icca10b59ade3079b2da5c21bacf465c8fb474b07
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3290887
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45483}