RELAND: [recovery] Forgiveness and restart count for Karte
This CL is a reland of the CL below.

https://chromium-review.googlesource.com/c/infra/infra/+/3290665

That CL was reverted using this CL:

https://chromium-review.googlesource.com/c/infra/infra/+/3298752

As described in b:207492932, the original CL iterated over a map and
used that to populate an array, which meant that the unit test in
question became non-deterministic.

BUG=b:206866884
TEST=make test

Change-Id: Ia75ebd426664460b8c02ea0699e9974cb2b765eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3298762
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45488}