Run auditwheel repair when creating manylinux wheels.
This removes the dependency on system libffi for the cffi wheel, and
also a bad dependency on zlib by opencv_python. Bump the patch versions
for these two wheels to cause new wheels to be built.

No other wheels appear to be affected.

We need to skip running auditwheel for a handful of wheels which are
platform-specific but contain no compiled code. Introduce a
skip_auditwheel parameter to SourceOrPrebuilt for this.

Bug: 1247674
Change-Id: I4d29cc518197900136f42ee56062c9ad981320f6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3295633
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#45492}