[Weetbix] Add methods to query re-clustering progress.
Provide separate methods for determining progress updating
re-clustering to use new algorithms and new rules.

These methods will be used to surface re-clustering progress on
the UI, and in bug-filing to determine whether analysis is
up-to-date (in general, vis-a-vis algorithms version, and for a
specific rule, vis-a-vis rules version).

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: Iec7b95892127be167f91a3fe17632238bebf0772
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3295635
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>
Cr-Commit-Position: refs/heads/main@{#45503}