[Weetbix] Fix flake in Rulescache tests.
The Initially Non-Empty Then Non-Empty test was flaky because there were
multiple rows with the same LastUpdated time. Line 49 of cache_test.go
expected sorted rules to be indexwise equiivalent if the sets were equal,
but this was not true, because the sort used was not deterministic.

BUG=1243174
TEST=go test infra/appengine/weetbix/...

Change-Id: Id0dff3b817e174c46837c9e9c824b38e7adf1669
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3301570
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#45567}