[recovery]: Check whether servod fails to detect CR50 due to low voltage.
This commit is a part of the PARIS project.

If the voltage values obtain from servod controls 'servo_dut_sbu1_mv'
and 'servo_dut_sbu2_mv' are below a specific threshold value, it
indicates that servo was unable to detect CR50 in the setup. This
commit adds a verifier that uses this logic to determine whether CR50
is connected.

Bug=b:198809511
TEST=build

Change-Id: I4620ba8dfa38885467828349a932026a611c9933
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3299393
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45572}