[cloudbuildhelper] Validate `notify` section in the YAML.
Previously it was completely pass-through from the YAML to
the JSON output. In the future CLs `cloudbuildhelper` will
learn to apply restrictions on what's allowed in the YAML,
and in particular we'll need to restrict allowed target Git
repos in `notify` section of the config. Thus `cloudbuildhelper`
needs to understand what is there.

R=iannucci@chromium.org

Change-Id: Ieec202cc71da43c62203e5b7c51097b125ef827e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3307610
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45602}