[recovery]: Include a newly added peripheral state to ufs converters.
This commit is a part of the PARIS project.

The peripheral states supported by UFS need to be supported in
PARIS. This includes the update to ufs converters.

In the past, a new peripheral state was added to UFS protobuf through
b/197684437. The current commit adds a support for this new state in
PARIS.

Bug=b:198809511
TEST=build

Change-Id: I949b87d5aa0ac6a540f4084708ca0260c78348b0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3311772
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#45730}