[pylint] Install pylint in correct directory
--target correctly installs pylint and all of its dependencies in the
specified directory and sets up a `bin` subdirectory. On the other hand,
using --root creates a directory structure like
`pylint/usr/local/home/me` where /usr/local/home/me is the value of
$HOME. I don't fully understand the meaning of --root and why it causes
this, but it's clearly wrong.

With --root:
  pylint
  └── usr
      └── local
	  └── google
	      └── home
		  └── olivernewman (empty directory)

With --target:
  pylint
  └── bin
      └── pylint (file)
  (other contents omitted)

I believe this was introduced in crrev.com/c/1976078 and deployed on
2020/01/07, so the pylint package has been broken for nearly 2 years.

Bug: 1275987
Change-Id: Iea62968e7e5bcec34d7f85c7cc54986a5dc12967
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3313442
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>
Cr-Commit-Position: refs/heads/main@{#45737}