[pylint] Propagate error if pylint doesn't exist
The pylint executable is not included in recent versions of the pylint
tricium function's CIPD package, but pylint_parser doesn't propagate
any error resulting from the executable not existing because:
1. It runs pylint as a Python script instead of as an executable, so
   `cmd.Start()` succeeds because the Python executable does exist.
2. It ignores all errors from `cmd.Wait()`, notably the `No such file or
   directory` error (with retcode 2) emitted by Python when the
   specified script does not exist.

I fixed this by checking if the pylint executable exists before trying
to run it, and adding error propagation.

This will help catch it much more quickly if the CIPD package becomes
broken again in the future.

Bug: 1275987
Change-Id: I87e307706abbb61e0a520ed43bb57b33505d74d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3312164
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45739}