[shivas] Set name of PARIS recovery task to constant.
Recoverylib (PARIS) explicitly exports a few constants for naming tasks,
including TaskNameRecovery and TaskNameDeploy. Use TaskNameRecovery for
the recovery task. Before this CL landed, the old name was "repair", not
"recovery". TaskNameRecovery is an opaque type that wraps the string
"recovery", so this CL *does* change the name of the task.

BUG=None
TEST=run repair on labstation via ./shivas repair -paris NAME_OF_LABSTATION and ./shivas repair NAME_OF_LABSTATION

Change-Id: I581ca39f5f0f47c8919dac2fd5fe2dda316c06c4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3311760
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#45794}