Fix exe failure checking logic in reportBootstrapFailure.
The condition for checking if the triggered exe failed was inverted.
This prevents the summary markdown from being updated in the case that
there is an error during the bootstrapping process and presumably would
lead to undesired modifications to the build proto when the exe was
successfully triggered.

Change-Id: I2368bf7af095019dc59f334913614bec82b589e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3318777
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#45843}