[crosskylabadmin] Add stub for new repair flow.
Add stubs for the new repair flow for labstations. I'm also adding the logic to
read from the CrOSSkylabAdmin config and use that to determine whether to use
the new flow or the old flow.

Issuing the buildbucket RPC calls necessary to actually use the new repair flow
is intentionally not implemented as part of this CL.

BUG=b:206044795
TEST=push to dev CrOSSKylabAdmin AND unittests

Change-Id: I36a3645109fd47d19e3f56c0679a5581a4456fe3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3309531
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#45846}