tricium pylint: use python3 version of pylint
Background: The pylint analyzer uses its own bundled
version of both python and pylint. Pylint is downloaded
and put in place before deployment via pip.

On my glinux workstation now, I can install the python3
version of pylint and not the python2 version, and we're
migrating to python3 in general now.

As far as I know, the python3 version with the same
rcfile should produce the same warnings as before.

Bug: 1275987
Change-Id: I3cae819cdf45ecd9b19f3ad2d426d988b8ca1ce7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3318778
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45849}