[pylint] Disable py3-specific warnings
It's not possible to fix these warnings without breaking compatibility
with python 2, which is bad because the recipe engine still requires
that all recipe code maintain compatibility with python 2 even if it
runs on python 3 in production.

These are all just cleanup suggestions anyway and not indicative of
bugs, so it's fine to ignore them.

Bug: 1275987
Change-Id: If0baa131a356960f31ce19bc8fef2ba84f4e7253
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3319156
Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45858}