[pylint] Use relative paths to input files
Previously pylint_parser would pass absolute paths to pylint, so pylint
would output absolute paths that would be passed back to Tricium only
for Tricium to barf because it (understandably) doesn't know how to
interpret absolute paths. For example,
ci.chromium.org/b/8828493201719799057 has absolute paths in its
`tricium` output property, so the comments it emitted were not added to
the CL by Tricium.

The workaround is to run pylint in the `inputDir` and pass it relative
paths, which will cause it to return relative paths. I tested this
locally to confirm that it causes pylint to print paths that are
relative to `inputDir`, rather than absolute.

Bug: 1275987
Change-Id: I199e36bed3fe3727ca86f20e4a65da8d3bc7f0a8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3321183
Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#45911}