[crosskylabadmin] Add config entry for UFS service.
Currently, CrOSSkylabAdmin does not talk directly to UFS via the UFS
client. Add an entry to the config so that we can contact UFS.

This CL is part of a collection of CLs (some still to be written) that
enables a slow rollout of repair for labstations. The same logic will
later be reused to incrementally roll out Paris for DUTs.

Inspecting the DUTs by issuing swarming RPC calls is an antipattern, so
for new functionality we will reach out to UFS.

BUG=b:206044795
TEST=make test

Change-Id: I45b6dedd090748188ad249671ae29058b1a01802
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3322098
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#45939}