[recovery]: Create verifier for power button.
This commit is a part of the PARIS project.

This commit adds a verifier to check that the 'pwr_button' signal
shows the power button has been released. When 'pwr_button' is stuck
at 'press, it commonly indicates that the ribbon cable is
disconnected.

Bug=b:198809511
TEST=build, manual test

Change-Id: I5df92ef217382f1fd8ef20ddda0974cddb5a59a0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3316838
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#45961}