Refactor the bootstrapper's main.go to provide better coverage.
This reorganizes the bootstrapper's main.go so that functionality
external to the binary is encapsulated behind functions so that more of
the internal logic can be tested.

This enables providing coverage of the logic for updating the build in
the event of failures: previously the conditions for setting the
ExeFailure tag and checking it were both inverted, which was leading to
the build summary being updated with the failure only if the
bootstrapped exe could not be executed. The ExeFailure tag was removed
in favor of just checking the underlying exception type as was
previously done for applying the ExeFailure tag. There are now test
cases to show that the build is updated appropriately.

Bug: 1210566
Change-Id: I567613214833094e2bcd0314ff3f2c0765648e9f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3319581
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#45963}